pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "soccer-prediction"
version = "0.1.0"
description = "Machine learning-based football match prediction system"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Bogdan", email = "bogdan@example.com"},
]
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.103.0",
    "uvicorn>=0.23.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.4.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[tool.setuptools]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--strict-markers"

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[project.urls]
Homepage = "https://github.com/user/soccer-prediction"
Repository = "https://github.com/user/soccer-prediction"

[project.scripts]
soccer-prediction = "src.api.main:main"